requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pandas==2.0.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
loguru==0.7.2
websockets==12.0
requests==2.31.0
